refactor/godClass-tornaUtil
